Move NetCDFMeteoManager to weather_reader mod
